Temporary hack to make AMD require(['renderjs']) work.
Thanks to Caleb James DeLisle.

A better fix in this case would be to cleanly allow RenderJS gadget
instanciation to wait for some events (require and any other JS code)
before deciding that a gadget is ready. Currently, RenderJS ready
function is triggered when all JS scripts defined in the HTML page are
loaded. One idea would be to extend the API to allow the gadget to
explicitely trigger the ready callback (but it will block the full
system if one gadget failed to load).